+++ java-jet/src/main/java/com/hbm/devices/jet/Peer.java
@@ -1,5 +1,5 @@
 package com.hbm.devices.jet;
 
 public interface Peer {
-	public void connect(double timeoutMs);
+	public void connect(int timeoutMs);
 }
